refactor(app): extract attention center UI from main.ts
Keep the desktop entry under the 19172-line split gate by moving attention rendering into attentionCenterPanel.ts.
