+++ .githooks/pre-commit-whitelist.ps1
@@ -193,6 +193,7 @@ $whitelistPatterns = @(
     'winsmux-app/src/firstRunOnboarding.ts',
     'winsmux-app/src/agentVaultOrganize.ts',
     'winsmux-app/src/attentionCenter.ts',
+    'winsmux-app/src/attentionCenterPanel.ts',
     'winsmux-app/src/firstRunWizard.ts',
     'winsmux-app/src/workflowGate.ts',
     'winsmux-app/src/styles.css',
